sentinel_fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
cryptography
